+++ .github/workflows/checks.yml
@@ -52,12 +52,10 @@ jobs:
 
   # Runs the test suite under the Thread Sanitizer to validate the
   # `@unchecked Sendable` and `nonisolated(unsafe)` sites at runtime.
-  # Non-blocking for now, until we confirm it is stable.
   test-tsan:
     name: Test (Thread Sanitizer)
     runs-on: macos-26
     if: ${{ !github.event.pull_request.draft }}
-    continue-on-error: true
     env:
       scheme: ${{ 'Readium-Package' }}
