+++ CHANGELOG.md
@@ -7,18 +7,14 @@ All notable changes to this project will be documented in this file. Take a look
 
 ### Changed
 
-#### All
-
-* The toolkit is migrated to Swift 6 with strict concurrency checking. All packages compile in the Swift 6 language mode, except `ReadiumInternal` which will follow in a future release. See [the migration guide](docs/Migration%20Guide.md).
-* The toolkit adopts the `NonisolatedNonsendingByDefault` (SE-0461), `InferIsolatedConformances` and `MemberImportVisibility` upcoming Swift features. Async APIs now run on the caller's actor, except CPU-heavy operations (decryption, parsing, search) which are marked `@concurrent` and stay off-actor. See [the migration guide](docs/Migration%20Guide.md).
-* The CocoaPods podspecs now declare `swift_version` 6.0.
+* The toolkit is migrated to Swift 6 with strict concurrency checking. All packages compile in the Swift 6 language mode. See [the migration guide](docs/Migration%20Guide.md).
+    * The toolkit adopts the `NonisolatedNonsendingByDefault` (SE-0461), `InferIsolatedConformances` and `MemberImportVisibility` upcoming Swift features.
 
 #### Shared
 
 * OPDS models (`Feed`, `Group`, `Facet`, `OpdsMetadata`) are now structs with value semantics.
 * `Publication`, `Resource`, `Container` and related types are now `Sendable`. Custom implementations of `Resource`, `Container`, `HTTPClient` or `PublicationService` must be `Sendable` too.
 * `Resource.stream()` now cooperates with task cancellation: the built-in resources fail with `ReadError.cancelled` when the surrounding task is cancelled, and custom implementations are expected to do the same.
-* `Weak` now requires the wrapped type to be `Sendable`.
 
 #### Navigator
