+++ Sources/Shared/Publication/Services/Cover/GeneratedCoverService.swift
@@ -70,7 +70,6 @@ public final class GeneratedCoverService: CoverService, Sendable {
             .success(ResourceProperties())
         }
 
-        /// `@concurrent` keeps the PNG encoding off the caller's actor.
         @concurrent func stream(range: Range<UInt64>?, consume: @escaping @Sendable (Data) -> Void) async -> ReadResult<Void> {
             await cover().flatMap {
                 guard let data = $0.pngData() else {
