+++ Tests/SharedTests/Toolkit/CancellableTasksTests.swift
@@ -4,7 +4,7 @@
 //  available in the top-level LICENSE file of the project.
 //
 
-import ReadiumInternal
+import ReadiumShared
 import Testing
 
 struct CancellableTasksTests {
